requirements.txt
numpy
numba
matplotlib
python-opencv
scikit-learn
lmfit
Pillow
